Fix second formatting in timer periods and total time; improve naming and refactor complexity labels
## PR Description:
This PR fixes the lack of proper second formatting for `timerPeriod.periodTime` and `widget.timer.totalTime`, fixes typos in variable names, improves function names and removes `simple` and `complex` from the `formatPeriodName()` function, by isolating it to a separate function.

## Problem:
The previous formatter only showed minutes, not seconds. This caused timers with periods whose total `periodTimes` were under 60 seconds to display as 0 minutes. Additionally, for the `totalTime` of an entire timer that included a period with a `periodTime` under 60 seconds, it would show invalid data by displaying only the minutes without the seconds, which could be misleading.

## Solution:
The formatting has been updated, to take into account if there are only seconds, minutes or both possible in `periodTime` or `totalTime`. It has been implemented as a reusable function `formatSeconds()`, and takes in an argument of type `int` that it then calculates seconds and/or minutes from it, returning them as a `String`.

## Changes:
- Introduced `formatSeconds()`: A reusable function to properly format seconds and/or minutes to a `String` from an argument of type `int`.
- Introduced `capitalize()`: A reusable function to capitalize the first letter of the string passed as an argument.
- Renamed `replacePeriodName()` to `formatPeriodName()`.
- Renamed `timerPeroid` to `timerPeriod`.
- Extracted `simple` and `complex` from the switch instruction in `formatPeriodName()` to a separate function `capitalize()`: This change was made, because `simple` and `complex` aren't part of `periodName`.