Add gradient for preset timer titles in timer picker columns
## PR Description:
This PR adds unique gradients to the titles of preset timers in 'timerPicker' columns. It adds a new function to achieve this.

## Problem:
Previously, all timers in the `timerPicker` columns, regardless wether they were the user's or preset, would all have the default, hardcoded Pomodoro gradient as their `fontColor`. This worsened the ability to easily identify timers based on the color of their title.

## Solution:
A new function - `getTimerTitleGradient(String timerName)` has been added, that takes in a `String` that returns a list of colors (`List<Color>`). Inside it has a `switch` instruction, that compares the passed String to any timers with preassigned gradients. Instead of hardcoding the `mainTextGradient` to the Pomodoro colors, it now calls a function that takes `timer.name` allowing for more gradient variations based on timer name.

## Changes:
- Added getTimerTitleGradient(String timerName) which returns a List<Color> based on a switch-case comparison.
- Replaced the hardcoded Pomodoro gradient with a call to the new function in relevant areas.
- Introduced specific gradients for 20-20-20 and a fallback/default gradient for all others.