Update timer column on box values change
## Commit Description:
Adds the necessary ability for the timer picker column to update it's contents once the timer type's (user, preset) box values have changed (for example, on delete).
To achieve this, the column holding all the TimerPickerColumnButtons was wrapped within a ValueListenableBuilder, that listened to any changes regarding the value of the selected Hive box.