Fix 'Empty' text not functioning properly
## Commit Description:
Moves the ValueListenableBuilder from the column to the very top of the widget tree, to account for the logic when there is no timers in the chosen box.

## Problem:
When deleting the last timer in the chosen box, the 'Empty' text would not show properly. The text 'empty' is supposed to be shown when there are no timers in the chosen box.

## Solution:
Moved the ValueListenableBuilder from the column to the very top, where the logic that decides whether to show the text `Empty` or the timers occurs. Now instead of rebuilding only the column, it rebuilds the whole widget, which allows the 'Empty' text to function properly.