Add TimerControlButton widget and fix TimerRunner layout issues
## Commit Description:
Adds the layout for `TimerControlButton`, a widget meant to control the running timer in `TimerRunner`.
Also adds the `TimerControlButton` and `ReturnButton` to the bottom column inside the `TimerRunner` layout. Changes around the margin insets to adjust the layout so it looks more like the prototype. Decreased the font size and faded out the header text (name of the running timer) for the top `PageHeader`.