Add ability to control countdown timer
## Commit Description:
Adds a TimerControlButton, giving the user the ability to start, pause or resume active countdown timers.
Apart from adding the TimerControlButton widget, it also adds a new callback to CircularTimer - onTimerControllerReady, that will be called everytime the timerAnimationController is assigned a new periodDuration (basically everytime a new period begins). Also added a new variable inside TimerRunner's state - timerAnimationController that will be assigned inside the callback of CircularTimer. The timerAnimationController inside TimerRunner is then passed to TimerControlButton's property of the same name, which allows the control button to control the countdown timer.