Update ReturnButton to use `popCurrentPage()` for `defaultClickLogic`
## Commit Description:
Replaces the custom function for `defaultClickLogic` with the standardized `popCurrentPage()`.
They are different in the way that the previous logic also forwarded the `pageAnimationController` (responsible for slide in, out animations) and waited until it finished before popping the page, while `popCurrentPage()` just pops the page, nothing else. This resulted in a prettty long wait after clicking the button, which could negatively impact the users by slowing them down with animations.
Adjusts the comments of `ReturnButton`'s class.