Remove unnecessary `src` folder
## Commit Description:
Takes out all of `src`'s contents and moves them all to the top of the directory, removing the `src` folder.
This was done so anyone looking through this repository doesn't have to dig to find `lib` and other important files, and so it isn't required to set the directory to `src` to actually `flutter run`.