Add comments to `timer_structure.dart`
## Commit Description:
Adds missing comments and documentation in `timer_structure.dart`. Before, it was a massive file with no explanation of how it worked or why certain enums or classes were combined. This commit fixes this issue by adding the missing comments.
Also improves comments in `timer_formatters` in it's `formatSeconds` function by explaining why the 's' was added to the end of the formatted string. Only 1 line, and not worth enough to have it's own commit, so thats why it's combined here.