Update generic_buttons comments
## Commit Description:
This a pretty miniscule commit, because of phenomenon that during the phase of creating buttons, I finally understood (Thanks To Christ), how commenting my code was crucial and such a great practice. The only file lacking in comments was `SquareButton`s `headers.dart` , which was completely barren in terms of comments.
This commit adds the necessary logic explanation required for maintainers to understand this file.