Replace all custom dividers with `BasicDivider`
## Commit Description:
Adds the `BasicDivider` in other widgets, replacing the old more custom and repeating dividers with the standardized `BasicDivider`.
Renames the `BasicDivider`'s file name from `divider` to `basic_divider` for clarity.
Adds a new property to `BasicDivider` - `fontAlpha`, allows to change the opacity of the font color without having to pass a color with `withValues()`.