Add leave timer popup
## Commit Description:
Adds a popup widget, that when trying to leave an already started timer, will display a confirmation message asking the user if they really want to leave the timer.
In `TimerRunner` added a `popUpContent` variable, `pushPopupContent` function and changed the Return Button's onPressed field to a function that first rebuilds the widget (to have the latest version of the `TimerControlButton`'s `currentAction`, and then if the `currentAction` isn't `start` then it shows the `LeaveTimerRunnerPopup` to make sure the user doesn't leave the timer by accident.