Remove every page's existing `top_bar`
## Commit Description:
Removes all `top_bar`s from pages' structures, because they would move with the page when it slid and for a brief second when switching pages, the user had no way to minimize or close the application.
Also removes an unnecessary `popup.dart` import inside `TimerRunner`, and improves the `top_bar` class comments, notifying that it is only to be used within `window_initializer` (inside `yourbreakApp`).