Add missing `UserStatsStructure` fields and it's `init`'s initialization methods
## Commit Description:
### `UserStatsStructure`
Adds missing `lastStreakDate` field in `UserStatsStructure` right under `loginStreak`. Without this, it would't have been possible to track the last time the user logged in, having no real way to actually track the streak.
Also adds `extendLoginStreak` as a method, that compares `DateTime.now()` against `lastStreakDate`, and if 1 day has passed then increment the streak by 1, but if more than 1 days have passed, then reset the streak to 1.
Also regenerates it's adapter.
### `UserStatsStructureInit`
Adds a new method called `_putDefaultUser()`, that puts a new entry with a key named `default_user` inside `user_stats` box, with every stat reset and the `lastStreakDate` set to the year 1900 (explain in a second).
Inside it's`init`, checks the `hasRunBefore` flag, and if it's false then call `_putDefaultUser()`.
Finally, get the `default_user` from `user_stats` box and call it's `extendLoginStreak()`.
WHY THE YEAR 1900: `extendLoginStreak` compares the current time with the `lastStreaDate`. If inside `_putDefaultUser` we set it's `lastStreakDate` to `DateTime.now()`, and soon after we call `extendLoginStreak` which also runs `DateTime.now()`, then there won't be a big enough time difference to count the streak.
Why 1900? Because it's really far back, and since `_putDefaultUser` is ran at the first launch of the app, that means that the user's streak is 1. If `extendLoginStreak` catches that the difference between now and `lastStreakDate` is bigger than 1, then it will reset the streak back to 1. Since 1900 was 125 years ago, it guarantees more than 1 day of difference.